requirements.txt
numpy==1.24.0
torch==1.13.0
matplotlib==3.5.2
hydra-core==1.3.2
jax==0.4.6
flax==0.6.3
optax==0.1.4